Add flash messages
This commit added a flash-message.blade.php file which contains messages
which will be showed when something like trying to delete a record
associated with a child record will be not possible.

+++ app/Http/Controllers/CampusController.php
@@ -97,6 +97,10 @@ class CampusController extends Controller
      */
     public function destroy(Campus $campus)
     {
+        if ($campus->memberships->count()) {
+            return back()->with('warning', 'No es posible borrar "' . $campus->name . '", se encuentra asociado a alguna membresía');
+        }
+
         $campus->delete();
 
         return redirect('/campus');
